updated attribute_type script

+++ gnowsys-ndf/gnowsys_ndf/ndf/views/ajax_views.py
@@ -193,36 +193,40 @@ def make_module_set(request, group_id):
             _id = request.GET.get("_id","")
             if _id:
                 node = collection.Node.one({'_id':ObjectId(_id)})
-                list_of_collection.append(node._id)
-                dict = {}
-                dict['id'] = unicode(node._id)
-                dict['version_no'] = hm_obj.get_current_version(node)
-                if node.collection_set:
-                    dict['collection'] = get_module_set_list(node)          #gives the list of collection with proper hierarchy as they are
+                if node:
+                    list_of_collection.append(node._id)
+                    dict = {}
+                    dict['id'] = unicode(node._id)
+                    dict['version_no'] = hm_obj.get_current_version(node)
+                    if node.collection_set:
+                        dict['collection'] = get_module_set_list(node)  #gives the list of collection with proper hierarchy as they are
            
-                #creating new Gsystem object and assining data of collection object
-                gsystem_obj = collection.GSystem()
-                gsystem_obj.name = unicode(node.name)
-                gsystem_obj.content = unicode(node.content)
-                gsystem_obj.member_of.append(GST_MODULE._id)
-                gsystem_obj.group_set.append(ObjectId(group_id))
-                # if usrname not in gsystem_obj.group_set:        
-                #     gsystem_obj.group_set.append(int(usrname))
-                gsystem_obj.created_by = int(request.user.id)
-                gsystem_obj.module_set.append(dict)
-                module_set_md5 = hashlib.md5(str(gsystem_obj.module_set)).hexdigest() #get module_set's md5
-
-                check =check_module_exits(module_set_md5)          #checking module already exits or not
-                if(check == 'True'):
-                    return HttpResponse("This module already Exists")
-                else:
-                    gsystem_obj.save()
-                    check1 = sotore_md5_module_set(gsystem_obj._id, module_set_md5)
-                    if (check1 == 'True'):
-                        return HttpResponse("module succesfull created")
+                    #creating new Gsystem object and assining data of collection object
+                    gsystem_obj = collection.GSystem()
+                    gsystem_obj.name = unicode(node.name)
+                    gsystem_obj.content = unicode(node.content)
+                    gsystem_obj.member_of.append(GST_MODULE._id)
+                    gsystem_obj.group_set.append(ObjectId(group_id))
+                    # if usrname not in gsystem_obj.group_set:        
+                    #     gsystem_obj.group_set.append(int(usrname))
+                    gsystem_obj.created_by = int(request.user.id)
+                    gsystem_obj.module_set.append(dict)
+                    module_set_md5 = hashlib.md5(str(gsystem_obj.module_set)).hexdigest() #get module_set's md5
+                    
+                    check =check_module_exits(module_set_md5)          #checking module already exits or not
+                    if(check == 'True'):
+                        return HttpResponse("This module already Exists")
                     else:
-                        gsystem_obj.delete()
-                        return HttpResponse("Error Occured while checking modules existance'")
+                        gsystem_obj.save()
+                        create_relation_of_module(node._id, gsystem_obj._id)
+                        check1 = sotore_md5_module_set(gsystem_obj._id, module_set_md5)
+                        if (check1 == 'True'):
+                            return HttpResponse("module succesfull created")
+                        else:
+                            gsystem_obj.delete()
+                            return HttpResponse("Error Occured while storing md5 of object in attribute'")
+                else:
+                    return HttpResponse("Object not present corresponds to this id")
             else:
                 return HttpResponse("Not a valid id passed")
         except Exception as e:
@@ -246,8 +250,15 @@ def sotore_md5_module_set(object_id,module_set_md5):
     else:
         print "Run 'python manage.py filldb' commanad to create AttributeType 'module_set_md5' "
         return 'False'
+
 #-- under construction    
-#def create_relation_module(object_id):
+def create_relation_of_module(subject_id, right_subject_id):
+    rt_has_module = collection.Node.one({'_type':'RelationType', 'name':'has_module'})
+    relation = collection.GRelation()                         #instance of GRelation class
+    relation.relation_type = rt_has_module
+    relation.right_subject = right_subject_id
+    relaton.subject = subject_id
+    relation.save()
     
 
 def check_module_exits(module_set_md5):
